+++ zingo-nymserverd/src/lib.rs
@@ -3,5 +3,4 @@
 #![warn(missing_docs)]
 #![forbid(unsafe_code)]
 
-pub mod proxy;
 pub mod server;
