+++ zingo-nymwalletd/Cargo.toml
@@ -1,15 +1,19 @@
 [package]
-name = "zingo-proxyd"
+name = "zingo-nymwalletd"
 version = "0.1.0"
 edition = "2021"
 
 [[bin]]
-name = "zingoproxyd"
-path = "src/bin/zingoproxyd.rs"
+name = "zingonymwalletd"
+path = "src/bin/nymwalletd.rs"
 
 [lib]
-name = "zingoproxylib"
+name = "nymwalletlib"
 path = "src/lib.rs"
+
+[features]
+default = ["nym_wallet"]
+nym_wallet = []
 
 [dependencies]
 zingo-rpc = { path = "../zingo-rpc" }
@@ -21,3 +25,13 @@ tonic = "0.10.2"
 zcash_client_backend = { git = "https://github.com/zingolabs/librustzcash.git", branch = "nym_integration", features = ["lightwalletd-tonic"] }
 # zingo-netutils = { git = "https://github.com/aloearev/zingolib.git", branch = "cratify_netutils" }
 zingo-netutils = { git = "https://github.com/zingolabs/zingolib.git", branch = "nym_integration" }
+prost = "0.12"
+bytes = "1.1"
+tokio-socks = "0.5"
+nym-sdk = { git = "https://github.com/nymtech/nym", branch = "master" }
+nym-sphinx-addressing = { git = "https://github.com/nymtech/nym", branch = "master" }
+nym-validator-client = { git = "https://github.com/nymtech/nym", branch = "master" }
+nym-bin-common = { git = "https://github.com/nymtech/nym", branch = "master" }
+nym-sphinx-anonymous-replies = { git = "https://github.com/nymtech/nym", branch = "master" }
+ctrlc = "3.2.1"
+
