+++ zingo-nymwalletd/zcash.conf
@@ -1,0 +1,11 @@
+## Default zingolib zcash.conf for regtest mode ##
+testnet=1
+
+txindex=1
+insightexplorer=1
+experimentalfeatures=1
+rpcuser=xxxxxx
+rpcpassword=xxxxxx
+rpcport=18232
+rpcallowip=127.0.0.1
+
