+++ zingo-nymwalletd/zebrad.toml
@@ -1,0 +1,2 @@
+network.network = "Testnet"
+rpc.listen_addr = '127.0.0.1:18232'
